# app_next/src/lib/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tank Battle API",
    "description": "API for Tank Battle - a multiplayer tank game.\n\n**Authentication:**\n- **Public API**: Use `X-API-Key` header with a valid API key (generate via `POST /api/me/api-keys` with session auth).\n- **Session API**: Requires a valid NextAuth session (cookie-based).",
    "version": "1.1.0",
    "contact": {
      "name": "Tank Battle API Support"
    }
  },
  "servers": [
    {
      "url": "https://localhost:8443",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key generated from your profile. Format: `pk_<uuid>`. Generate keys via `POST /api/me/api-keys` (requires session auth)."
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth session cookie. Obtained by signing in via the web application."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Invalid API key" }
        }
      },
...
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "My CLI Key" },
          "key": { "type": "string", "example": "pk_abc123def456", "description": "Only returned on creation" },
          "isActive": { "type": "boolean", "example": true },
          "expiresAt": { "type": "string", "format": "date-time", "nullable": true },
          "lastUsedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/me/api-keys": {
      "get": {
        "tags": ["Me"],
        "summary": "List API keys",
        "description": "Get all API keys for the authenticated user.",
        "security": [{ "SessionAuth": [] }],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/ApiKey" } } } } } }
          },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "tags": ["Me"],
        "summary": "Create API key",
        "description": "Generate a new API key for accessing the public API.",
        "security": [{ "SessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "My CLI Key" },
                  "expiresInDays": { "type": "integer", "example": 30, "description": "Optional expiration in days" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key created",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" }, "data": { "$ref": "#/components/schemas/ApiKey" } } } } }
          },
          "400": { "description": "Missing name", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["Me"],
        "summary": "Delete API key",
        "description": "Revoke and delete an API key.",
        "security": [{ "SessionAuth": [] }],
        "parameters": [
          { "name": "id", "in": "query", "required": true, "schema": { "type": "integer" }, "description": "API key ID to delete" }
        ],
        "responses": {
          "200": {
            "description": "API key deleted",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "message": { "type": "string" } } } } }
          },
          "400": { "description": "Invalid ID", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/friends/request": {
      "post": {
        "tags": ["Friends"],
        "summary": "Send friend request",
        "description": "Send a friend request to another user by username.",
        "security": [{ "SessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username"],
                "properties": {
                  "username": { "type": "string", "example": "TankPro" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Friend request sent", "content": { "application/json": { "schema": { "type": "object", "properties": { "id": { "type": "integer" }, "status": { "type": "string" } } } } } },
          "400": { "description": "Already friends or request exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "User not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/friends/accept": {
      "post": {
        "tags": ["Friends"],
        "summary": "Accept friend request",
        "description": "Accept a pending friend request.",
        "security": [{ "SessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["friendshipId"],
                "properties": {
                  "friendshipId": { "type": "integer", "example": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Friend request accepted", "content": { "application/json": { "schema": { "type": "object", "properties": { "friendship": { "type": "object" }, "conversation": { "type": "object" } } } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Request not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/friends/deny": {
      "post": {
        "tags": ["Friends"],
        "summary": "Deny friend request",
        "description": "Decline a pending friend request.",
        "security": [{ "SessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["friendshipId"],
                "properties": {
                  "friendshipId": { "type": "integer", "example": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Friend request denied", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Request not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/game/results": {
      "post": {
        "tags": ["Game"],
        "summary": "Submit game results",
        "description": "Submit the authenticated user's game result after a multiplayer match.",
        "security": [{ "SessionAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["leaderboard"],
                "properties": {
                  "leaderboard": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "userId": { "type": "integer" },
                        "username": { "type": "string" },
                        "playerNumber": { "type": "integer" },
                        "score": { "type": "integer" }
                      }
                    }
                  },
                  "gameMode": { "type": "integer", "description": "Number of players (2, 3, or 4)" },
                  "duration": { "type": "integer", "description": "Game duration in seconds" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Results saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } },
          "400": { "description": "Invalid data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
...
}